Remove hard-coded filepaths, use --train_dir, --val_dir, --log_dir, --checkpoint_dir
